refactor(auth): move pure instance-token primitives to @epicenter/auth
The instance bearer's two pure pieces (the token generator and the boot entropy gate) have no server coupling: they touch only Web Crypto. Move generateInstanceToken, assertStrongToken, and MIN_INSTANCE_TOKEN_CHARS from @epicenter/server into @epicenter/auth so a token can be minted and validated without the server graph (the CLI and the self-host scripts depend on @epicenter/auth, not @epicenter/server).

The verifier side (createInstanceTokenResolver, verifyEnvToken, INSTANCE_PRINCIPAL, VerifyToken) needs AuthUser/ResolveUser, so it stays in @epicenter/server. Repoint the self-host gen-token script and Bun entry; the generator/gate test cases move to packages/auth alongside the code.
