+++ apps/self-host/package.json
@@ -16,6 +16,7 @@
 		"dev:bun": "bun --watch server.ts"
 	},
 	"dependencies": {
+		"@epicenter/auth": "workspace:*",
 		"@epicenter/constants": "workspace:*",
 		"@epicenter/server": "workspace:*",
 		"arktype": "catalog:",
