+++ packages/auth/src/index.ts
@@ -29,6 +29,16 @@ export {
 	type InstanceSetting,
 	loadInstanceSetting,
 } from './instance-setting.js';
+// The pure pieces of the single-partition instance bearer (self-host; ADR-0073):
+// `generateInstanceToken` mints a strong token (`gen-token`), `assertStrongToken`
+// is the boot entropy gate. They live here (not `@epicenter/server`) so a token
+// can be generated and validated without the server graph. The verifier side
+// (`createInstanceTokenResolver`, `verifyEnvToken`) stays in `@epicenter/server`.
+export {
+	assertStrongToken,
+	generateInstanceToken,
+	MIN_INSTANCE_TOKEN_CHARS,
+} from './instance-token.js';
 export {
 	type CreateInstanceTokenAuthConfig,
 	createInstanceTokenAuth,
