+++ packages/server/src/auth/instance-token.test.ts
@@ -1,23 +1,19 @@
 /**
- * Instance-token credential unit tests (ADR-0073).
+ * Instance-token verifier unit tests (ADR-0073).
  *
- * Pins the four pieces of the instance bearer path: the resolver maps an exact,
+ * Pins the verifier side of the instance bearer path: the resolver maps an exact,
  * constant-time bearer match to the named principal and everything else to
- * `InvalidToken`; the entropy gate fails closed on a missing / short / hand-typed
- * token and passes a strong one; and the generator's output always satisfies the
- * gate (the lockstep that lets the boot error point operators at `gen-token`). The
- * surface wrappers' HTTP/WebSocket shaping is covered in `require-auth.test.ts`.
+ * `InvalidToken`. The surface wrappers' HTTP/WebSocket shaping is covered in
+ * `require-auth.test.ts`; the pure generator + entropy gate (`generateInstanceToken`
+ * / `assertStrongToken`) live in `@epicenter/auth` and are tested there.
  */
 
 import { expect, test } from 'bun:test';
 import type { Context } from 'hono';
 import type { Env } from '../types.js';
 import {
-	assertStrongToken,
 	createInstanceTokenResolver,
-	generateInstanceToken,
 	INSTANCE_PRINCIPAL,
-	MIN_INSTANCE_TOKEN_CHARS,
 	verifyEnvToken,
 } from './instance-token.js';
 
@@ -68,38 +64,4 @@ test('rejects a missing Authorization header with InvalidToken', async () => {
 test('rejects a non-bearer scheme with InvalidToken', async () => {
 	const { error } = await resolve(contextWithAuthorization(`Basic ${TOKEN}`));
 	expect(error?.name).toBe('InvalidToken');
-});
-
-test('generateInstanceToken emits a 256-bit base64url token that clears the gate', () => {
-	const token = generateInstanceToken();
-	expect(token).toMatch(/^[A-Za-z0-9_-]+$/); // base64url, no padding
-	expect(token.length).toBe(43); // 32 bytes -> 43 base64url chars
-	expect(token.length).toBeGreaterThanOrEqual(MIN_INSTANCE_TOKEN_CHARS);
-	// Lockstep: the generator must always satisfy the gate, or the boot error
-	// would reject a token the operator just generated.
-	expect(assertStrongToken(token)).toBe(token);
-	expect(generateInstanceToken()).not.toBe(token); // fresh randomness each call
-});
-
-test('assertStrongToken fails closed on a missing or empty token', () => {
-	expect(() => assertStrongToken(undefined)).toThrow(/not set/);
-	expect(() => assertStrongToken('')).toThrow(/not set/);
-	expect(() => assertStrongToken('   ')).toThrow(/not set/);
-});
-
-test('assertStrongToken fails closed on a short token', () => {
-	expect(() => assertStrongToken('letmein')).toThrow(/too weak/);
-	expect(() =>
-		assertStrongToken('a'.repeat(MIN_INSTANCE_TOKEN_CHARS - 1)),
-	).toThrow(/too weak/);
-});
-
-test('assertStrongToken fails closed on a passphrase (spaces / control chars)', () => {
-	expect(() =>
-		assertStrongToken('correct horse battery staple correct horse'),
-	).toThrow(/URL-safe/);
-});
-
-test('assertStrongToken returns the trimmed token for a strong value', () => {
-	expect(assertStrongToken(`  ${TOKEN}  `)).toBe(TOKEN);
 });
